give the random token helper its own home
